Completely strip out repo-add symlink compatibility
If this compatibility is desirable, either first create the repo with
repo-add or manually move repose's created repository and make the
symlinks yourself.

This idea has been fraught with enough corner cases that make me doubt
having it as a toggleable behaviour is not a good idea and I'd rather
not make the symlinks by default since they're there in the first place
due to limitations of bsdtar.