Fix repose failing to pick up new signatures
Closes #21 (for real, this time).

There are to problems. The first was that I was lazily loading the
PGPSIG field. Which completely defeated the timestamp checking since it
was happening far too late, way after I've decided the repo should be
rebuild. Then, it turns out, the logic for comparing mtime was
backwards so this.