Remove the dead filter fields from the db struct
This field was here so we could support automatically detecting the
compression to use when updating the repo. This field isn't being used
because the feature isn't implemented.

While this is feature that may appear, I don't think the field is useful
anyways because we throw away the database object before we could read
that field out anyways.