link signatures from pool to root, not just packages
Works for both symlinks and reflinks. In no case does it do any true
error handling for either one; it simply avoids creating symlinks if
source doesn't exist in the pool, and reflinks if it can't open either
end of the link.

This is on the grounds that the detached signature files aren't strictly
necessary for any purpose (as the database contains individual package
signatures) and therefore there's no need to complain if they don't
exist, and little need to complain if it can't create them.